% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koreaAddressParse.R
\encoding{UTF8}
\name{KAParse}
\alias{KAParse}
\title{KAParse: Parse Korea address as per areas}
\usage{
KAParse(dfDirtyJuso)
}
\arguments{
\item{dfInput}{Input data frame}

\item{columnName}{column name of input data frame, which will be parsed}
}
\value{
data frame is reaturned and it have administrative areas as new columns
}
\description{
This function is to parse Korea address string as per administrative areas.
}
\details{
KAParse read only the column which is pointed by columnName parameter, and add new columns for administrative areas such as SiDoMyeong, SiGunGuMyeong. If we can't find appropriated administrative area, we mark 'x' to its corresponding column.
}
\examples{
library(KoreaAddressAPI)

dirtyJuso = c('攼㸶댘尸ｆ꽱尸ㅶ댘尸愼㸶攼㸶뵾尸┕愼㹦攼㸶댘尸→댘媤경댘尸ㅶ뙹媤戼㸳 \n\t   攼㸶댘媤득댘尸愼㸶攼㸶뙹媤겡길댘媤戼㸸 | |   攼㸶댘尸€뵾媤룡댘媤당동',
'攼㸶댘尸ｆ꽱尸ㅶ댘尸愼㸶攼㸶뵾尸┕愼㹦攼㸶댘尸→댘媤경댘尸ㅶ뙹媤戼㸳 攼㸶댘尸ζ샎媤멩댘媤는길댘媤戼㸸 攼㸶뙹媤겠당동')
dfDirtyJuso = data.frame(dirtyJuso)
dfDirtyJuso

dfCleaned = KoreaAddressParse(dfDirtyJuso)
dfCleaned
}
